requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autoflake==2.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
discord==2.3.1
discord.py==2.3.1
freezegun==1.2.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.27.8
pyfiglet==0.8.post1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
